chore: move gcloud into utils
